\name{grp.mat}
\alias{grp.mat}
\title{grp.mat}
\usage{
  grp.mat(mat.ce, mat.l, mat.w)
}
\arguments{
  \item{mat.ce}{}

  \item{mat.l}{}

  \item{mat.w}{}
}
\value{
  single matrix
}
\description{
  catch/effort and sample data from three matrix to single
  matrix form
}
\author{
  Pierre Kleiber
}

